Warehouse-to-sales point transfer feature added
Introduced new view, controller actions, and service methods for transferring products from warehouses to sales points. Updated view models and resource files for improved localization and UI consistency. Enhanced Production Index with new transfer option.

+++ Inventar.Web.ViewModels/Logistics/TransferProductItemViewModel.cs
@@ -11,6 +11,6 @@ namespace Inventar.Web.ViewModels.Logistics
         public Guid ProductId { get; set; }
         public string ProductName { get; set; } = null!;
         public double AvailableQuantity { get; set; }
-        public double TransferQuantity { get; set; }
+        public int TransferQuantity { get; set; }
     }
 }
